Add BDD-Security template; cleanup
Added a new template for bdd-security.  Will require us to add an
environment variable to our servers though, so not totally happy
about it

Renamed template-deploy to template-deploy-ansible so that we have
multiple types of deploys.  This was mostly because we originally
thought we'd only ever do Ansible deployments, but the need to support
Fabric is still strong enough that we should at least be flexible

Added in log rotation for all our templates