1. Adds slightly modified versions of RestApiScriptRunner and RestApiJobManagement from https://github.com/sheehan/job-dsl-gradle-example 2. renames all example jobs for consistency and absence of redundancy
Details:

1. Script runners:
  Irina had originally included these in the internal
jenkins-automation-private repo, for use with recovering the "Mother
Seed Job".
  I think this is so useful for a local development workflow that I'm
bringing it into jenkins-automation
  It will then be usable by any repo that uses jenkins-automation.
  The idea is to use this as part of a local development workflow for
job-dsl jobs that mitigates the slow-and-annoying git workflow we have
ot use now

2. Example job renaming

These jobs are already in an `example-jobs` folder.
I see no need to have them have `example` or `sample` in the name.
In addition, they should be named consistently, not the island of misfit
names that they were prior to this commit.