Updating to latest jobs-dsl-core plugin v1.57 (#95)
* Updating to latest jobs-dsl-core plugin v1.57

In the latest release the runDslEngine() was deprecated. This
commit makes the required API changes.

* Merging crumb support from sheenah/job-dsl-grable-example

This commit merges the crumb support from the master repo.

* Fixing assert due to invalid merge

The previous commit incorrectly copied over the line,
ContentType.TEXT, when this was pulled from a different library.
This caused a runtime assert. The fix is to revert this code to the
correct object.