Remove 'presetSpecified' param since it appears to have no effect
I can't find any jobs where this is set, but more importantly, it looks like
it is always 'false' in the jobs' config.xml, even when a preset has been
specified (I assume 'specified' means a preset that is not the default preset).

If you look at the plugin code, it doesn't look like they are ever setting
'presetSpecified' to 'true'. I can also find nothing in the docs that describes
what this setting does. Until we figure it out, seems best to not allow it to
be specified as a param to the builder.

Also, there was no mention of this param in the CheckmarxSecurityJobBuilder's
groovydoc, so that seems like another indicator that it probably isn't being
set to 'true' by anyone -- how would they know what it did?