Include 'vulnerabilityThresholdResult' so that threshold-based failures work
Right now, exceeding a threshold won't fail the job properly, even when you
have configured it so that you think it should. The problem is that a chunk
of xml is missing from the job config, and without it, the failure won't work.
This adds that chunk of xml back, so that now when a threshold is exceeded,
and the job is configured to have 'vulnerabilityThresholdEnabled', the failure
will actually happen as expected.

How did I discover this? I made a JAC checkmarx job, inspected the config.xml,
then did a 'save' on the Configure screen in the UI. I saw that this stuff
got added as a side-effect of the save, suggesting that it was missing to
begin with.