Factor Checkmarx 'configure' block out of the builder for reusability
We're moving the `configure` block that adds Checkmarx scanning
out of the `CheckmarxSecurityJobBuilder` class's `build` method and
into an independent block (aka 'util', for now) for better reusability.

We needed to do this so that we could add Checkmarx scanning in the context
of the seed job creation -- where we would not have wanted to use the Checkmarx
builder class. And in retrospect, there is some consensus that this builder
probably should never have existed, and should just have been a block the
whole time, but we'll preserve the builder for now to minimize disruption.

However there _will_ be some disruption to existing consumers of
`CheckmarxSecurityJobBuilder`, since the constructor params have been changed:
instead of passing in all the Checkmarx config options as args to the
constructor, we're passing them in as a `Map`. This seems cleaner and it also
meant that we centralized where the defaults for the Checkmarx config
options are being set. Unfortunately this change means that existing jobs
will need to be updated to work with the new constructor setup.