BDDSecurityBuilder: change gradle test to gradle clean test (#129)
* Copy `baseUrl` into the JENKINS_URL environment variable for the REST
runner

We use the JENKINS_URL environment variable in several jobs.
When seed jobs run in the Jenkins runtime, Jenkins automatically
supplies this variable. However, when creating jobs using the REST
runner, the variable does not exist. This change ensures that
environment variable is available by setting it to the required
`baseUrl` parameter

* Untangles SauceConnectJobBuilder from JsJobBuilder

This removes the addition of a default NodeJS installation and removes a
default archiveArtifacts.

Both of these changes are desirable, as they do not make sense as
defaults on this builder

Addresses DEV-2478

* Adds clean. Fixes DEV-2530