Remove FlowJobBuilder and associated examples
The Jenkins Build Flow plugin is no longer maintained and has several
known security vulnerabilities. With the release of JEP-200 earlier
this year, it's even more urgent to abandon support for Build Flow
jobs, since the plugin is not being updated to get JEP-200
compatibility. For these reasons, it makes sense for us to now remove
the FlowJobBuilder class and related bits of example code.