Update job dsl (#145)
* Update job-dsl dependency to latest

We're now running the latest job-dsl-plugin on our Jenkins servers. This
gets our automation helpers up to that version

* Update job-dsl to latest; fix test to accomodate

This updates to the latest version of job-dsl.

I also brought over the syntax tests that Irina built for cf.gov, and adapted them for our example jobs. This is a vast improvement over the original test.

* fix formatting