Add credentials binding util (#146)
* Update job-dsl dependency to latest

We're now running the latest job-dsl-plugin on our Jenkins servers. This
gets our automation helpers up to that version

* Update job-dsl to latest; fix test to accomodate

This updates to the latest version of job-dsl.

I also brought over the syntax tests that Irina built for cf.gov, and adapted them for our example jobs. This is a vast improvement over the original test.

* Adds a utility to configure usernamePassword and amazonWebServices credentials without needing job-dsl dynamic methods

The problem with job-dsl's dynamic methods is that they do not work in the context of `gradlew test` and `gradlew rest`. This addition enables us to use clean usernamePassword and amazonWebServices credentials in our jobs and still retain gradle-based unit and rest test workflows.

Improves APPOPS-444 considerably.