Add option to email build log using BaseJobBuilder
This commit adds the ability to optionally attach job build logs to
notification emails when using the BaseJobBuilder builder class. This
is intended to make it easier to debug job failures after the Jenkins
log history has been purged.

The CommonUtils.addExtendedEmail helper now has a new attachBuildLog
option that defaults to false. This option is also added to BJB, which
passes it on to addExtendedEmail.

See Jenkins Job DSL plugin documentation for extendedEmail here:
https://jenkinsci.github.io/job-dsl-plugin/#method/javaposse.jobdsl.dsl.helpers.publisher.PublisherContext.extendedEmail